+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## 1.7 (work in progress)
 
-- [x] Optimized for iOS 15, iPadOS 15, and macOS 12
-- [x] Add Location Button (iOS 15, macOS 12)
+- [x] **Optimized for iOS 15, iPadOS 15, and macOS 12**
+- [x] **Loans widget (iOS 14, macOS 11)**
+- [x] **Loans notifications (1 week before due date)**
+- [x] List reservations
+- [x] Location Button (iOS 15, macOS 12)
 - [x] Filter agenda by library (iOS 14, macOS 11)
-- [x] Add Drag and Drop support (News, Agenda)
+- [x] Drag and Drop support (News, Agenda)
 - [x] Update libraries opening hours
+- [x] Improve agenda parsing
 - [x] Fix show current location on library maps